bug: 无法完成 drop 和 drop all，停止开发这两个功能

+++ src/main/java/top/guoziyang/mydb/backend/dm/pageCache/PageCacheImpl.java
@@ -55,7 +55,7 @@ public class PageCacheImpl extends AbstractCache<Page> implements PageCache {
     }
 
     public Page getPage(int pgno) throws Exception {
-        return get((long)pgno);
+        return get(pgno);
     }
 
     /**
